Итоговый commit

+++ database/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Table, Column, Integer, String, MetaData, ForeignKey
+from sqlalchemy import Table, Column, Integer, String, MetaData, BigInteger
 
 metadata_obj = MetaData()
 
@@ -6,7 +6,7 @@ metadata_obj = MetaData()
 users_table = Table(
     "users",
     metadata_obj,
-    Column("id", Integer, primary_key=True),
+    Column("id", BigInteger, primary_key=True),
     Column("username", String),
     Column("userphone", String),
     Column("usertgid", Integer),
@@ -16,7 +16,7 @@ users_table = Table(
 orders_table = Table(
     "orders",
     metadata_obj,
-    Column("id", Integer, primary_key=True),
+    Column("id", BigInteger, primary_key=True),
     Column("client_name", String),
     Column("client_phone", String),
     Column("client_tgid", String),
